chore(integration): Move sign-in-or-up-test into single file
